restructure codebase
