+++ packages/abell-examples/.eslintrc.js
@@ -1,5 +1,6 @@
 module.exports = {
   rules: {
-    '@typescript-eslint/no-var-requires': 0
+    '@typescript-eslint/no-var-requires': 0,
+    'spaced-comment': 0
   }
 };
