--- packages/abell/src/abell.ts
@@ -1,5 +1,0 @@
-declare module '*.abell' {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  export const html: (props?: any) => string;
-  export default html;
-}
